# api/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: quang.huynh@gotit.vn
  description: Technical document APIs for API Version 4.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Sample API
  version: 4.0.0
servers:
- description: API BIZ Version 4 Staging
  url: https://api-biz-stg.gotit.vn
...
paths:
  /api/v4.0/brands:
    get:
      description: Retrieve a list of brands
      operationId: getListOfBrands
      parameters:
      - description: Authorization
        explode: false
        in: header
        name: X-GI-Authorization
        required: true
        schema:
          example: API key GotIt provided
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BRANDS_RESPONSE'
          description: Success
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTP_BAD_REQUEST'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTP_UNAUTHORIZED'
          description: Unauthenticated
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTP_NOT_FOUND'
          description: Request not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTP_INTERNAL_SERVER_ERROR'
          description: System error
      summary: Get list of brands
      tags:
      - Brands
  /api/v4.0/brands/{id}:
    get:
      description: Returns brand detail data master
      operationId: getDetailOfBrand
      parameters:
      - description: Authorization
        explode: false
        in: header
        name: X-GI-Authorization
        required: true
        schema:
          example: API key GotIt provided
          type: string
        style: simple
      - description: Brand Id
        explode: false
        in: path
        name: id
        required: true
        schema:
          example: 46
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BRAND_DETAIL_RESPONSE'
          description: Success
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTP_BAD_REQUEST'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTP_UNAUTHORIZED'
          description: Unauthenticated
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTP_NOT_FOUND'
          description: Request not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTP_INTERNAL_SERVER_ERROR'
          description: System error
      summary: Get brand detail
      tags:
      - Brands
  /api/v4.0/brands/brandsByProducts:
    get:
      description: Returns brand detail data master
      operationId: brandsByProducts
      parameters:
      - description: Authorization
        explode: false
        in: header
        name: X-GI-Authorization
        required: true
        schema:
          example: API key GotIt provided
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BRAND_DETAIL_RESPONSE'
          description: Success
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTP_BAD_REQUEST'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTP_UNAUTHORIZED'
          description: Unauthenticated
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTP_NOT_FOUND'
          description: Request not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTP_INTERNAL_SERVER_ERROR'
          description: System error
      summary: Get brand by product
      tags:
      - Brands
  /api/v4.0/categories:
    get:
      description: Returns all categories data master
      operationId: getListOfCategories
      parameters:
      - description: Authorization
        explode: false
        in: header
        name: X-GI-Authorization
        required: true
        schema:
          example: API key GotIt provided
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CATEGORIES_RESPONSE'
          description: Success
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTP_BAD_REQUEST'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTP_UNAUTHORIZED'
          description: Unauthenticated
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTP_NOT_FOUND'
          description: Request not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTP_INTERNAL_SERVER_ERROR'
          description: System error
      summary: Get lists category
      tags:
      - Categories
  /api/v4.0/categories/categoriesByProducts:
    get:
      description: Returns all categories data master
      operationId: getCategoryByProduct
      parameters:
      - description: Authorization
        explode: false
        in: header
        name: X-GI-Authorization
        required: true
        schema:
          example: API key GotIt provided
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CATEGORIES_RESPONSE'
          description: Success
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTP_BAD_REQUEST'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTP_UNAUTHORIZED'
          description: Unauthenticated
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTP_NOT_FOUND'
          description: Request not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTP_INTERNAL_SERVER_ERROR'
          description: System error
      summary: Get category by product
      tags:
      - Categories
  /api/v4.0/products:
    get:
      description: Returns all products master data
      operationId: getListOfProducts
      parameters:
      - description: Authorization
        explode: false
        in: header
        name: X-GI-Authorization
        required: true
        schema:
          example: API key GotIt provided
          type: string
        style: simple
      - description: Page
        explode: true
        in: query
        name: page
        required: true
        schema:
          example: 1
          type: integer
        style: form
      - description: Page Size
        explode: true
        in: query
        name: pageSize
        required: true
        schema:
          example: 200
          type: integer
        style: form
      - description: Min price
        explode: true
        in: query
        name: minPrice
        required: false
        schema:
          example: 1000
          type: integer
        style: form
      - description: Max price
        explode: true
        in: query
        name: maxPrice
        required: false
        schema:
          example: 10000000
          type: integer
        style: form
      - description: Stores
        explode: true
        in: query
        name: isExcludeStoreListInfo
        required: false
        schema:
          example: true
          type: boolean
        style: form
      - description: Store Page
        explode: true
        in: query
        name: storeListPage
        required: false
        schema:
          example: 1
          type: integer
        style: form
      - description: Store Page Size
        explode: true
        in: query
        name: storeListPageSize
        required: false
        schema:
          example: 200
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PRODUCTS_RESPONSE'
          description: Success
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTP_BAD_REQUEST'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTP_UNAUTHORIZED'
          description: Unauthenticated
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTP_NOT_FOUND'
          description: Request not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTP_INTERNAL_SERVER_ERROR'
          description: System error
      summary: Get all products master data
      tags:
      - Products
  /api/v4.0/products/{id}:
    get:
      description: Returns product detail data
      operationId: getProductDetail
      parameters:
      - description: Authorization
        explode: false
        in: header
        name: X-GI-Authorization
        required: true
        schema:
          example: API key GotIt provided
          type: string
        style: simple
      - description: Product ID
        explode: false
        in: path
        name: id
        required: true
        schema:
          example: 1541
          type: integer
        style: simple
      - description: Exclude store list information
        explode: true
        in: query
        name: isExcludeStoreListInfo
        required: false
        schema:
          example: true
          type: boolean
        style: form
      - description: Store Page
        explode: true
        in: query
        name: storeListPage
        required: false
        schema:
          example: 1
          type: integer
        style: form
      - description: Store Page Size
        explode: true
        in: query
        name: storeListPageSize
        required: false
        schema:
          example: 200
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PRODUCT_DETAIL_RESPONSE'
          description: Success
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTP_BAD_REQUEST'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTP_UNAUTHORIZED'
          description: Unauthenticated
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTP_NOT_FOUND'
          description: Request not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTP_INTERNAL_SERVER_ERROR'
          description: System error
      summary: Get product detail data
      tags:
      - Products
  /api/v4.0/products/{id}/stores:
    get:
      description: Returns all stores of this product
      operationId: getStoresOfProduct
      parameters:
      - description: Authorization
        explode: false
        in: header
        name: X-GI-Authorization
        required: true
        schema:
          example: API key GotIt provided
          type: string
        style: simple
      - description: Product ID
        explode: false
        in: path
        name: id
        required: true
        schema:
          example: 1541
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/STORES_RESPONSE'
          description: Success
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTP_BAD_REQUEST'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTP_UNAUTHORIZED'
          description: Unauthenticated
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTP_NOT_FOUND'
          description: Request not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTP_INTERNAL_SERVER_ERROR'
          description: System error
      summary: Get stores of this product
      tags:
      - Products
  /api/v4.0/vouchers/multiple/status/{refId}:
    get:
      description: Check voucher status
      operationId: checkVoucher
      parameters:
      - description: Authorization
        explode: false
        in: header
        name: X-GI-Authorization
        required: true
        schema:
          example: API key GotIt provided
          type: string
        style: simple
      - description: Transaction RefId
        explode: false
        in: path
        name: refId
        required: true
        schema:
          example: voucher_0711_102
          type: string
        style: simple
      - description: Exclude store list information
        explode: true
        in: query
        name: isExcludeStoreListInfo
        required: false
        schema:
          example: true
          type: boolean
        style: form
      - description: Store Page
        explode: true
        in: query
        name: storeListPage
        required: false
        schema:
          example: 1
          type: integer
        style: form
      - description: Store Page Size
        explode: true
        in: query
        name: storeListPageSize
        required: false
        schema:
          example: 200
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VOUCHER_CHECK_RESPONSE'
          description: Success
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTP_BAD_REQUEST'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTP_UNAUTHORIZED'
          description: Unauthenticated
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTP_NOT_FOUND'
          description: Request not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTP_INTERNAL_SERVER_ERROR'
          description: System error
      summary: Check voucher status
      tags:
      - Voucher Status
  /api/v4.0/vouchers/v:
    post:
      description: Create voucher link v
      operationId: createVoucherLinkV
      parameters:
      - description: Authorization
        explode: false
        in: header
        name: X-GI-Authorization
        required: true
        schema:
          example: API key GotIt provided
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/REQUEST_CREATE_VOUCHER_LINK_V'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VOUCHER_V_RESPONSE'
          description: Success
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTP_BAD_REQUEST'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTP_UNAUTHORIZED'
          description: Unauthenticated
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTP_NOT_FOUND'
          description: Request not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTP_INTERNAL_SERVER_ERROR'
          description: System error
      summary: Create voucher link v
      tags:
      - Voucher
...
components:
...
  schemas:
...
    HTTP_BAD_REQUEST:
      example:
        data: "[]"
        error: HTTP_BAD_REQUEST
        message: Bad Request
        status: HTTP_BAD_REQUEST
        statusCode: 400
      properties:
        status:
          example: HTTP_BAD_REQUEST
          type: string
        statusCode:
          example: 400
          type: integer
        error:
          example: HTTP_BAD_REQUEST
          type: string
        message:
          example: Bad Request
          type: string
        data:
          example: "[]"
          items: {}
          type: array
      type: object
    HTTP_UNAUTHORIZED:
      example:
        data: "[]"
        error: HTTP_UNAUTHORIZED
        message: Unauthorized
        status: HTTP_UNAUTHORIZED
        statusCode: 401
      properties:
        status:
          example: HTTP_UNAUTHORIZED
          type: string
        statusCode:
          example: 401
          type: integer
        error:
          example: HTTP_UNAUTHORIZED
          type: string
        message:
          example: Unauthorized
          type: string
        data:
          example: "[]"
          items: {}
          type: array
      type: object
    HTTP_NOT_FOUND:
      example:
        data: "[]"
        error: HTTP_NOT_FOUND
        message: Not Found
        status: HTTP_NOT_FOUND
        statusCode: 404
      properties:
        status:
          example: HTTP_NOT_FOUND
          type: string
        statusCode:
          example: 404
          type: integer
        error:
          example: HTTP_NOT_FOUND
          type: string
        message:
          example: Not Found
          type: string
        data:
          example: "[]"
          items: {}
          type: array
      type: object
...